Move temperature alert logic to Content.Shared
