+++ Content.Server/Temperature/Systems/TemperatureSystem.cs
@@ -1,26 +1,18 @@
 using Content.Server.Atmos.EntitySystems;
-using Content.Shared.Alert;
-using Content.Shared.Body.Components;
 using Content.Shared.Temperature.Components;
 using Content.Shared.Temperature.Systems;
-using Robust.Shared.Prototypes;
 
 namespace Content.Server.Temperature.Systems;
 
 public sealed class TemperatureSystem : SharedTemperatureSystem
 {
-    [Dependency] private readonly AlertsSystem _alerts = default!;
     [Dependency] private readonly AtmosphereSystem _atmosphere = default!;
-
-    [ValidatePrototypeId<AlertCategoryPrototype>]
-    public const string TemperatureAlertCategory = "Temperature";
 
     public override void Initialize()
     {
         base.Initialize();
 
         SubscribeLocalEvent<TemperatureComponent, AtmosExposedUpdateEvent>(OnAtmosExposedUpdate);
-        SubscribeLocalEvent<AlertsComponent, OnTemperatureChangeEvent>(ServerAlert);
     }
 
     private void OnAtmosExposedUpdate(EntityUid uid, TemperatureComponent temperature,
@@ -37,62 +29,5 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
         var heat = temperatureDelta * (airHeatCapacity * heatCapacity /
                                        (airHeatCapacity + heatCapacity));
         ChangeHeat(uid, heat * temperature.AtmosTemperatureTransferEfficiency, temperature: temperature);
-    }
-
-    private void ServerAlert(EntityUid uid, AlertsComponent status, OnTemperatureChangeEvent args)
-    {
-        ProtoId<AlertPrototype> type;
-        float threshold;
-        float idealTemp;
-
-        if (!TryComp<TemperatureComponent>(uid, out var temperature))
-        {
-            _alerts.ClearAlertCategory(uid, TemperatureAlertCategory);
-            return;
-        }
-
-        if (TryComp<ThermalRegulatorComponent>(uid, out var regulator) &&
-            regulator.NormalBodyTemperature > temperature.ColdDamageThreshold &&
-            regulator.NormalBodyTemperature < temperature.HeatDamageThreshold)
-        {
-            idealTemp = regulator.NormalBodyTemperature;
-        }
-        else
-        {
-            idealTemp = (temperature.ColdDamageThreshold + temperature.HeatDamageThreshold) / 2;
-        }
-
-        if (args.CurrentTemperature <= idealTemp)
-        {
-            type = temperature.ColdAlert;
-            threshold = temperature.ColdDamageThreshold;
-        }
-        else
-        {
-            type = temperature.HotAlert;
-            threshold = temperature.HeatDamageThreshold;
-        }
-
-        // Calculates a scale where 1.0 is the ideal temperature and 0.0 is where temperature damage begins
-        // The cold and hot scales will differ in their range if the ideal temperature is not exactly halfway between the thresholds
-        var tempScale = (args.CurrentTemperature - threshold) / (idealTemp - threshold);
-        switch (tempScale)
-        {
-            case <= 0f:
-                _alerts.ShowAlert(uid, type, 3);
-                break;
-
-            case <= 0.4f:
-                _alerts.ShowAlert(uid, type, 2);
-                break;
-
-            case <= 0.66f:
-                _alerts.ShowAlert(uid, type, 1);
-                break;
-
-            case > 0.66f:
-                _alerts.ClearAlertCategory(uid, TemperatureAlertCategory);
-                break;
-        }
     }
 }
